+++ cartridges/int_klaviyo_core/cartridge/scripts/klaviyo/eventData/addedToCart.js
@@ -33,6 +33,7 @@ function getData(basket) {
             var lineItem = basketItems[itemIndex];
             var currentProductID = lineItem.productID;
             var basketProduct = ProductMgr.getProduct(currentProductID);
+
             if (!basketProduct) {
                 throw new Error('Product with ID [' + currentProductID + '] not found');
             }
@@ -67,9 +68,10 @@ function getData(basket) {
                     primaryCategory           : primaryCategory
                 };
 
-                if (!basketProduct.master && 'masterProduct' in basketProduct) {
-                    currentLineItem.masterProductID = basketProduct.masterProduct.ID;
+                if (basketProduct.variant) {
+                    currentLineItem.masterProductID = klaviyoUtils.getParentProductId(basketProduct);
                 }
+
 
                 var priceData = klaviyoUtils.priceCheck(lineItem, basketProduct);
                 currentLineItem.price = priceData.purchasePrice;